Stop overwriting the users login shell
Before we loaded bash with no profile to make it quick, hide MOTDs etc.
We actually do want aliases and co to work though so we want to get rid
of this behavior.  Further we no longer spawn shells in --command more
anymore which is also a good thing.

Note polysh has special hacks for bash to recognize the prompt and
overwrite the PS1 which may not work for other shells at the moment.