Apply patches/type-hints.diff
This fixes most type issues in polysh and adds type hints.  We still
need to disabled mypys strict optional checking for now though:
https://mypy.readthedocs.io/en/latest/kinds_of_types.html#no-strict-optional

Further we're using a variable in asyncore which is not yet documented
in the upstream typeshed type hints.

Sadly we need to drop python3.4 support to support type hints.